Add ImageDecoderRegistry test
