+++ Sources/Core/Decoding/ImageDecoderRegistry.swift
@@ -1,0 +1,98 @@
+// The MIT License (MIT)
+//
+// Copyright (c) 2015-2021 Alexander Grebenyuk (github.com/kean).
+
+import Foundation
+
+/// An image decoder which supports automatically registering in the decoder register.
+public protocol ImageDecoderRegistering: ImageDecoding {
+    /// Returns non-nil if the decoder can be used to decode the given data.
+    ///
+    /// - parameter data: The same data is going to be delivered to decoder via
+    /// `decode(_:)` method. The same instance of the decoder is going to be used.
+    init?(data: Data, context: ImageDecodingContext)
+
+    /// Returns non-nil if the decoder can be used to progressively decode the
+    /// given partially downloaded data.
+    ///
+    /// - parameter data: The first and the next data chunks are going to be
+    /// delivered to the decoder via `decodePartiallyDownloadedData(_:)` method.
+    init?(partiallyDownloadedData data: Data, context: ImageDecodingContext)
+}
+
+public extension ImageDecoderRegistering {
+    /// The default implementation which simply returns `nil` (no progressive
+    /// decoding available).
+    init?(partiallyDownloadedData data: Data, context: ImageDecodingContext) {
+        return nil
+    }
+}
+
+// MARK: - ImageDecoderRegistry
+
+/// A registry of image codecs.
+public final class ImageDecoderRegistry {
+    /// A shared registry.
+    public static let shared = ImageDecoderRegistry()
+
+    private struct Match {
+        let closure: (ImageDecodingContext) -> ImageDecoding?
+    }
+
+    private var matches = [Match]()
+
+    public init() {
+        self.register(ImageDecoders.Default.self)
+        self.register(ImageDecoders.Video.self)
+    }
+
+    /// Returns a decoder which matches the given context.
+    public func decoder(for context: ImageDecodingContext) -> ImageDecoding? {
+        for match in matches {
+            if let decoder = match.closure(context) {
+                return decoder
+            }
+        }
+        return nil
+    }
+
+    // MARK: - Registering
+
+    /// Registers the given decoder.
+    public func register<Decoder: ImageDecoderRegistering>(_ decoder: Decoder.Type) {
+        register { context in
+            if context.isCompleted {
+                return decoder.init(data: context.data, context: context)
+            } else {
+                return decoder.init(partiallyDownloadedData: context.data, context: context)
+            }
+        }
+    }
+
+    /// Registers a decoder to be used in a given decoding context. The closure
+    /// is going to be executed before all other already registered closures.
+    public func register(_ match: @escaping (ImageDecodingContext) -> ImageDecoding?) {
+        matches.insert(Match(closure: match), at: 0)
+    }
+
+    /// Removes all registered decoders.
+    public func clear() {
+        matches = []
+    }
+}
+
+/// Image decoding context used when selecting which decoder to use.
+public struct ImageDecodingContext {
+    public let request: ImageRequest
+    public let data: Data
+    /// Returns `true` if the download was completed.
+    public let isCompleted: Bool
+    public let urlResponse: URLResponse?
+
+    public init(request: ImageRequest, data: Data, isCompleted: Bool, urlResponse: URLResponse?) {
+        self.request = request
+        self.data = data
+        self.isCompleted = isCompleted
+        self.urlResponse = urlResponse
+    }
+}
