+++ Sources/Core/Decoding/ImageDecoders+Video.swift
@@ -2,7 +2,6 @@
 //
 // Copyright (c) 2015-2021 Alexander Grebenyuk (github.com/kean).
 
-
 #if !os(watchOS)
 
 import Foundation
